requirements.txt
flask==2.3.3
flask-socketio==5.3.6
python-socketio==5.9.0
eventlet==0.33.3
werkzeug==2.3.7
opencv-python-headless==4.8.1.78
mediapipe==0.10.7
numpy==1.24.3
pillow==10.0.1
scikit-learn==1.5.2
supabase==2.16.0
python-dotenv==1.0.0
gunicorn==21.2.0